+++ internet-shop_ct/RepositoryInterfaces/IUsersRepository.cs
@@ -4,7 +4,7 @@
     {
         T GetByUserCode(int user_code); //Получить пользователя по его коду
 
-        T GetByLogin(int phone_number, string password); //Получить пользователя по его данным (номер телефона и пароль)
+        T GetByLogin(decimal phone_number, string password); //Получить пользователя по его данным (номер телефона и пароль)
 
         T Add(T newUser); //Добавить пользователя
